LRPS Algorithm v.1.3.1
Links on MD modified, mainly image hyperlink url, now when click on image it opens the image instead of go to the main repo (like before). Also the alt texts on links in spanish MD was changed to spanish. The main links on the top block (about where to read the original/another language article) wash changed; only the Medium.com link (the original article link) is pending